refactor(names): extract validation helpers
